+++ core/src/main/scala/com/deciphernow/server/GMFabricServer.scala
@@ -28,6 +28,9 @@ import com.twitter.util.Time
   */
 abstract class GMFabricServer extends App {
 
+  override def allowUndefinedFlags: Boolean = com.deciphernow.server.config.flags.allowUndefinedFlags.apply
+  override def failfastOnFlagsNotParsed: Boolean = com.deciphernow.server.config.flags.failFastOnFlagsNotParsed.apply
+
   val log = Logger.get(getClass)
   var restServer : Option[GMFabricRestServer] = None
   var thriftServer : Option[GMFabricThriftServer] = None