+++ core/src/main/scala/com/deciphernow/server/support/DecryptorManager.scala
@@ -20,11 +20,10 @@ package com.deciphernow.server.support
 /**
   * Created by ghershfield on 6/8/16.
   */
-object DecryptorManager extends DecryptorManager
+object DecryptorManager extends DecryptorManager(Option(com.deciphernow.server.config.resources.decryptClass.apply))
 
-class DecryptorManager {
+class DecryptorManager(val className : Option[String] = Option(com.deciphernow.server.config.resources.decryptClass.apply)) {
 
-  lazy val className : Option[String] = com.deciphernow.server.config.resources.decryptClass.getWithDefault
   lazy val clazz: Class[_] = Class.forName(className.get)
   lazy val instance = clazz.newInstance().asInstanceOf[Decryptor]