+++ core/src/test/scala/com/deciphernow/server/support/MySillyDecryptor_1.scala
@@ -1,0 +1,12 @@
+package com.deciphernow.server.support
+
+class MySillyDecryptor_1 extends Decryptor {
+  override def decryptResource(string: String): String = {
+    if (string.startsWith("-->{")) {
+      "MySillyDecryptor_1"
+    }
+    else {
+      string
+    }
+  }
+}